refactor food entity.

+++ app/src/main/java/com/chs/your_body_profile/domain/model/FoodDetailInfo.kt
@@ -2,8 +2,8 @@ package com.chs.your_body_profile.domain.model
 
 data class FoodDetailInfo(
     val name: String,
-    val servingWeight: Float = 0f,
-    val calorie: Float = 0f,
+    val servingWeight: Int = 0,
+    val calorie: Int = 0,
     val carbohydrate: Float = 0f,
     val fat: Float = 0f,
     val protein: Float = 0f,
